+++ Sources/WalletConnect/JsonRpcHistory/JsonRpcRecord.swift
@@ -1,0 +1,17 @@
+
+
+import Foundation
+
+struct JsonRpcRecord: Codable {
+    let id: Int64
+    let topic: String
+    let chainId: String?
+    let request: Request
+    var response: JsonRpcResponseTypes?
+    
+    struct Request: Codable {
+        let method: ClientSynchJSONRPC.Method
+        let params: ClientSynchJSONRPC.Params
+    }
+}
+
